requirements.txt
aiohttp~=3.10.6
aiosqlite~=0.20.0
discord.py~=2.4.0
python-dotenv~=1.0.1
pydantic~=2.9.2
